Codex

Public

Codex.accuracy — Method

accuracy(trues, preds)::Number

The number of correct predictions in pred (by comparing true to prediction) divided by the total number of predictions.

Codex.apply — Method

apply(fns, obj)
apply(fns) -> Function
apply(fn::Function, nt::NamedTuple) -> NamedTuple

Apply function fn or functions fns to object. The functions are applied in order, unlike the behaviour of function composition. Also defines partial function. (For partial declarations in Base, see issue #35052 or endswith(suffix).)

Codex.cohens_d — Method

cohens_d(μ1, μ2, s)
cohens_d(n1, μ1, s1, n2, μ2, s2)
cohens_d(A::Array, B::Array)

Effect size according to Cohen's d for means μ1 and μ2, number of samples n1 and n2, and standard deviations s1 and s2 for respectively group 1 and 2.

Codex.dirparent — Method

dirparent(path)::String
dirparent(path, n)::String

Returns the parent or n-th parent directory for path, where path can be a file or directory.

dirparent("/a/b/c")

Codex.has_duplicates — Method

has_duplicates(A::AbstractArray)::Bool

Returns whether A contains duplicates.

Codex.map_by_df — Method

map_by_df(a::Array, df::DataFrame, from::Symbol, to::Symbol; missing=nothing)::Array

Return array A where all elements are mapped from U to V. Leaving all elements of A for which no match is found unchanged.

!! Note-to-self. This is dumb function. Use joins instead.

Codex.nrow_per_group — Method

nrow_per_group(df::DataFrame, group::Symbol; col1="group", col2="nrow")::DataFrame

Return the group name and the number of rows per group in df.

Codex.output — Method

output(f::Function) -> Output
output(cmd::Cmd) -> Output

Evaluates f of type f(out::String, err::String)::CmdRedirect or cmd::Cmd.

Codex.rescale — Method

rescale(a, a_l, a_u, b_l, b_u)::Number

Apply feature scaling to a from the range [a_l, a_u] to the range [b_l, b_u].

Codex.split_data — Method

split_data(df::DataFrame, group_col, data_col, groups)::Tuple

Return a NamedTuple containing data from data_col for each group in groups. The ordering of the data is the same as the ordering of groups.

Private