TransformExport

This module is used to transform the data exported from the backend.

Codex.TransformExport.process — Method

process(in_dir, out_dir; fns)

Processes the responses from the export folder, applies the functions fns and places the files at out_dir.

Codex.TransformExport.read_csv — Method

read_csv(path; delim)::DataFrame

Copies CSV at path into memory.

Codex.TransformExport.responses — Method

responses(dir::String)::Dict{String,DataFrame}

Return responses for an export folder such as "2020-08".

Codex.TransformExport.simplify — Method

simplify(df)::DataFrame

Renames id column after removing extraneous rows and columns, that is, removes empty rows and removes columns such as protocol_subscription_id, open_from and v2_1_timing.

Codex.TransformExport.substitute_names — Method

substitute_names(df, with::DataFrame)::DataFrame
substitute_names(with)::Function

Replaces person_ids by the first name as listed in with.