Codex

Public

Codex.add_missing — Method

add_missing(df::DataFrame, actual::ColumnIndex, expected::Array, by::ColumnIndex)::DataFrame

Add rows to ensure that for all elements in expected, the same element exists in actual for each unique element of by.

Codex.add_missing — Method

add_missing(df::DataFrame, actual::ColumnIndex, expected::Array)::DataFrame

Add rows to ensure that for all elements in expected, the same element exists in actual.

Codex.apply — Method

apply(fns, obj)
apply(fns)::Function

Apply functions fns to object obj. The functions are applied in order, unlike the behaviour of function composition. Also defines partial function. (For partial declarations in Base, see issue #35052 or endswith(suffix).)

Codex.categorical2simple — Method

categorical2simple(A::CategoricalArray)

Returns simple Julia collection for the CategoricalArray a.

Codex.dirparent — Method

dirparent(path)::String
dirparent(path, n)::String

Returns the parent or n-th parent directory for path, where path can be a file or directory.

dirparent("/a/b/c")

Codex.has_duplicates — Method

has_duplicates(A::AbstractArray)::Bool

Returns whether A contains duplicates.

Codex.map_by_df — Method

map_by_df(a::Array, df::DataFrame, from::Symbol, to::Symbol; missing=nothing)::Array

Map array a by using arrays from and to. Leaving all elements of a for which no match is found unchanged.

Codex.order_with — Method

order_with(df::DataFrame, orderings::Array{Ordering,1})::DataFrame

Sorts df by cols in the order given by orderings, see the documentation of DataFrames.sort. Let o be an ordering::Ordering. Enforces that the columns o.col of df are ordered according to o.newlevels if o.newlevels is not missing.

Codex.project_root — Method

project_root()::String

Returns root directory of the current Module. This is usually also the root of the Git repository.

Private