Codex

Public

Codex.accuracy — Method

accuracy(trues, preds)::Number

The number of correct predictions in pred (by comparing true to prediction) divided by the total number of predictions.

Codex.add_missing — Method

add_missing(df::DataFrame, actual::ColumnIndex, expected::Array, by::ColumnIndex)::DataFrame

Add rows to ensure that for all elements in expected, the same element exists in actual for each unique element of by.

Codex.add_missing — Method

add_missing(df::DataFrame, actual::ColumnIndex, expected::Array)::DataFrame

Add rows to ensure that for all elements in expected, the same element exists in actual.

Codex.apply — Method

apply(fns, obj)
apply(fns)::Function

Apply functions fns to object obj. The functions are applied in order, unlike the behaviour of function composition. Also defines partial function. (For partial declarations in Base, see issue #35052 or endswith(suffix).)

Codex.categorical2simple — Method

categorical2simple(A::CategoricalArray)

Returns simple Julia collection for the CategoricalArray a.

Codex.cohens_d — Method

cohens_d(μ1, μ2, s) 
cohens_d(n1, μ1, s1, n2, μ2, s2)
cohens_d(A::Array, B::Array)

Effect size according to Cohen's d for means μ1 and μ2, number of samples n1 and n2, and standard deviations s1 and s2 for respectively group 1 and 2.

Codex.dirparent — Method

dirparent(path)::String
dirparent(path, n)::String

Returns the parent or n-th parent directory for path, where path can be a file or directory.

dirparent("/a/b/c")

Codex.has_duplicates — Method

has_duplicates(A::AbstractArray)::Bool

Returns whether A contains duplicates.

Codex.map_by_df — Method

map_by_df(a::Array, df::DataFrame, from::Symbol, to::Symbol; missing=nothing)::Array

Map array a by using arrays from and to. Leaving all elements of a for which no match is found unchanged.

Codex.nrow_per_group — Method

nrow_per_group(df::DataFrame, group::Symbol; col1="group", col2="nrow")::DataFrame

Return the group name and the number of rows per group in df.

Codex.order_with — Method

order_with(df::DataFrame, orderings::Array{Ordering,1})::DataFrame

Sorts df by cols in the order given by orderings, see the documentation of DataFrames.sort. Let o be an ordering::Ordering. Enforces that the columns o.col of df are ordered according to o.newlevels if o.newlevels is not missing.

Codex.output — Method

output(f::Function) -> Output
output(cmd::Cmd) -> Output

Evaluates f of type f(out::String, err::String)::CmdRedirect or cmd::Cmd.

Codex.project_root — Method

project_root()::String

Returns root directory of the current Module. This is usually also the root of the Git repository.

Codex.rescale — Method

rescale(a, a_l, a_u, b_l, b_u)::Number

Apply feature scaling to a from the range [a_l, a_u] to the range [b_l, b_u].

Private