Questionnaires

Module for analysing the questionnaires.

Public

Private

Codex.Questionnaires.all_ids — Method

all_ids(responses_dir)::Vector{String}

Codex.Questionnaires.charlie2int — Method

charlie2int(q::Int, s::String)

Convert s to Int for question number q.

Codex.Questionnaires.dropouts — Method

dropouts(raw_dir::String)

Returns dropout data where all IDs are in the long identifier format. On 2021-07-19, returned a 197x7 dataset.

Codex.Questionnaires.fix_age — Method

fix_age(x::AbstractString)::Int

Return age after parsing x.

julia> fix_age = Codex.Questionnaires.fix_age;

julia> fix_age("27-jun-93")
26

julia> fix_age("1993")
26

julia> fix_age("23 jaar")
23

Codex.Questionnaires.get_hnd — Method

get_hnd(path::AbstractString)::DataFrame

Get HowNutsAreTheDutch data and select big five, age and more.

Codex.Questionnaires.join_questionnaires — Method

join_questionnaires(raw_dir::String, questionnaires::Array{String,1}, groups::Array{String,1})::DataFrame

Joines multiple questionnaires for the members of groups, where groups ensures that the joins do not remove rows which shouldn't be removed. For instance, given questionnaires "A" and "B" containing respectively columns :a1, :a2 and :b1, this method will return { group, id, A_a1, A_a2, B_b1 }.

Codex.Questionnaires.join_vo_questionnaires — Method

join_vo_questionnaires(raw_dir::String)::DataFrame

Combine information from multiple questionnaires to allow model fitting.

Codex.Questionnaires.replace_usernames — Method

replace_usernames(df::DataFrame, id_col::Symbol, id_username::DataFrame)

For every username in df[!, id_col], replace the name by the long identifier. The long identifier is necessary, because usernames where not anonymized in 2018.

Codex.Questionnaires.resilience2scores — Method

resilience2scores(df::DataFrame)

Return the score for questionnaire charlie.

Codex.Questionnaires.responses — Method

responses(data_dir::String, nato_name::String, group::String; measurement=999)::DataFrame

Responses for group group and measurement measurement, where group is one of graduates, operators, dropouts-medical or dropouts-non-medical. measurement is only used to split the 2018 data, for the later datasets it is ignored.

Example in 2021-07:

julia> dir = joinpath(ysf_raw, "2020-first");

julia> df = Codex.Questionnaires.responses(dir, "kilo", "dropouts");

julia> nrow(df)
25

julia> select(first(df, 5), Not([:id, :completed_at]))
5×3 DataFrame
 Row │ group     optimism  pessimism
     │ String    Int64     Int64
─────┼───────────────────────────────
   1 │ dropouts        10         10
   2 │ dropouts        13          6
   3 │ dropouts        10          9
   4 │ dropouts        12          6
   5 │ dropouts        12          5

Codex.Questionnaires.responses — Method

responses(data_dir::String, nato_name::String)::DataFrame

Responses for questionnaire nato_name as contained in directory data_dir. Returns a DataFrame with rows { id, r...} where id is a long identifier and not the one from the backend.

Codex.Questionnaires.self_efficacy2scores — Method

self_efficacy2scores(echo::DataFrame)::DataFrame

Return the scores for the echo questionnaire, which is about self-efficacy. None of the items appear to be reversed, so this method just returns the sum.

This function assumes questions [1, 14] which holds for all questionnaires in from 2018 to the time of writing (2021).

Codex.Questionnaires.unfinished_info — Method

unfinished_info(responses_dir; required::Union{Nothing,Vector{String}}=missing)::Vector{Unfinished}

Return the ids for which not all questionnaires in required have been filled in. When ismissing(required), take all the questionnaires in responses_dir.

Codex.Questionnaires.unfinished_questionnaires — Method

unfinished_questionnaires(responses_dir::String, id::String)

Return questionnaire names in responses_dir for which id is not in filled_out_by_id column. This assumes that unfinished questionnaires have a missing in the filled_out_by_id column.

Codex.Questionnaires.unify_demographics — Method

unify_demographics(df)

Returns an simplified and unified DataFrame which is the same for 2018, 2019 and 2020. It may throw out some data which we don't need at the time of writing.

Codex.Questionnaires.validate_dropouts — Method

validate_dropouts(df::DataFrame)

Validate "dropouts.csv".