Questionnaires

Module for analysing the questionnaires.

Public

Private

Codex.Questionnaires.dropouts — Method

dropouts(raw_dir::String)::DataFrame

Returns dropout data where all IDs are in the long identifier format.

Codex.Questionnaires.fix_age — Method

fix_age(x::AbstractString)::Int

Return age after parsing x.

julia> fix_age = Codex.Questionnaires.fix_age;

julia> fix_age("27-jun-93")
26

julia> fix_age("1993")
26

julia> fix_age("23 jaar")
23

Codex.Questionnaires.get_hnd — Method

get_hnd(path::AbstractString)::DataFrame

Get HowNutsAreTheDutch data and select big five, age and more.

Codex.Questionnaires.join_dropout_questionnaires — Method

join_dropout_questionnaires(raw_dir::String)::DataFrame

Combine information from multiple questionnaires to allow model fitting.

Codex.Questionnaires.join_questionnaires — Method

join_questionnaires(raw_dir::String, questionnaires::Array{String,1}, groups::Array{String,1})::DataFrame

Joines multiple questionnaires for the members of groups, where groups ensures that the joins do not remove rows which shouldn't be removed. For instance, given questionnaires "A" and "B" containing respectively columns :a1, :a2 and :b1, this method will return { group, id, A_a1, A_a2, B_b1 }.

Codex.Questionnaires.plot_domain_density — Method

plot_domain_density(df::DataFrame, y::Symbol)

Density plot for age and domain y on DataFrame df containing group.

Codex.Questionnaires.responses — Method

responses(data_dir::String, nato_name::String, group::String; measurement=999)::DataFrame

Responses for group group and measurement measurement, where group is one of graduates, operators, dropouts-medical or dropouts-non-medical. measurement is only used to split the 2018 data, for the later datasets it is ignored.

Codex.Questionnaires.responses — Method

responses(data_dir::String, nato_name::String)::DataFrame

Responses for questionnaire nato_name as contained in directory data_dir. Returns a DataFrame with rows { id, r...} where id is a long identifier and not the one from the backend.

Codex.Questionnaires.self_efficacy2scores — Method

self_efficacy2scores(echo::DataFrame)::DataFrame

Return the scores for the echo questionnaire, which is about self-efficacy. None of the items appear to be reversed, so this method just returns the sum.

This function assumes questions [1, 14] which holds for all questionnaires in from 2018 to the time of writing (2021).

Codex.Questionnaires.unify_demographics — Method

unify_demographics(df)

Returns an simplified and unified DataFrame which is the same for 2018, 2019 and 2020. It may throw out some data which we don't need at the time of writing.