Questionnaires

Module for analysing the questionnaires.

Public

Private

Codex.Questionnaires.dropouts — Method

dropouts(raw_dir::String)::DataFrame

Returns dropout data where all IDs are in the long identifier format.

Codex.Questionnaires.get_hnd — Method

get_hnd(path::AbstractString)::DataFrame

Get HowNutsAreTheDutch data and select big five, age and more.

Codex.Questionnaires.plot_domain_density — Method

plot_domain_density(df::DataFrame, y::Symbol)

Density plot for age and domain y on DataFrame df containing group.

Codex.Questionnaires.responses — Method

responses(data_dir::String, nato_name::String, group::String; measurement=999)::DataFrame

Responses for group group and measurement measurement, where group is one of graduates, operators, dropouts-medical or dropouts-non-medical. measurement is only used to split the 2018 data, for the later datasets it is ignored.

Codex.Questionnaires.responses — Method

responses(data_dir::String, nato_name::String)::DataFrame

Responses for questionnaire nato_name as contained in directory data_dir. Returns a DataFrame with rows { id, r...} where id is a long identifier and not the one from the backend.