TransformExport

Module to transform the data exported from the backend.

Public

Codex.TransformExport.process — Method

process(in_dir, out_dir; fns)

Processes the responses from the export folder, applies the functions fns and places the files at out_dir.

Codex.TransformExport.read_csv — Method

read_csv(path; delim)::DataFrame

Copies CSV at path into memory.

Codex.TransformExport.responses — Method

responses(dir::String)::Dict{String,DataFrame}

Return responses for an export folder such as "2020-08".

Codex.TransformExport.rm_descriptions — Method

rm_descriptions(df)::DataFrame

Find responses containing a description, for example 6 (heel erg), and remove the description.

Codex.TransformExport.simplify — Method

simplify(df)::DataFrame

Renames id column after removing extraneous rows and columns, that is, removes empty rows and removes columns such as protocol_subscription_id, open_from and v2_1_timing.

Codex.TransformExport.split_datetime — Method

split_datetime(df::DataFrame, datetime_col::ColumnIndex)::DataFrame

Split the datetime column datetime_col into two columns, namely one for date and one for time.

Codex.TransformExport.substitute_names — Method

substitute_names(df, with::DataFrame)::DataFrame
substitute_names(with)::Function

Replaces person_ids by the first name as listed in with.

Private

Codex.TransformExport._contains_description — Method

_contains_description(col)::Bool

Return whether the column col contains descriptions.

Codex.TransformExport._description_regex — Method

_description_regex()

Return regex for matching a description such as 1 (lorem) or 2 <br /> (ipsum).

Codex.TransformExport._rm_description — Method

_rm_description(e::String)::String

Apply regex replace on element e.

Codex.TransformExport._rm_descriptions — Method

_rm_descriptions(col)::Array{Int,1}

Apply a regex replace and type conversion to all elements of the column col.

Codex.TransformExport.parsedatetime — Method

parsedatetime(str)::DateTime

Parse a date and time string from the export to a Julia DateTime object.