TransformExport

This module is used to transform the data exported from the backend.

Codex.TransformExport.responses — Method

responses(dir::String)::Dict{String,DataFrame}

Return responses for an export folder such as "2020-08".